+++ include/reactphysics3d/mathematics/Vector2.h
@@ -220,7 +220,7 @@ RP3D_FORCE_INLINE bool Vector2::isUnit() const {
 
 // Return true if the values are not NAN OR INF
 RP3D_FORCE_INLINE bool Vector2::isFinite() const {
-    return std::isfinite(x) && std::isfinite(y);
+    return std::isfinite((double)x) && std::isfinite((double)y);
 }
 
 // Return true if the vector is the zero vector
